Master Update

+++ src/server/game/Movement/Spline/MovementUtil.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2005-2011 MaNGOS <http://getmangos.com/>
+ * Copyright (C) 
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,11 +17,12 @@
  */
 
 #include "MoveSplineFlag.h"
-#include <cmath>
+#include <math.h>
+#include <string>
 
 namespace Movement
 {
-    float gravity = static_cast<float>(19.29110527038574);
+    double gravity = 19.29110527038574;
     UInt32Counter splineIdGen;
 
     /// Velocity bounds that makes fall speed limited
@@ -44,14 +45,14 @@ namespace Movement
             if (path_length >= terminal_safeFall_length)
                 time = (path_length - terminal_safeFall_length) / terminalSafefallVelocity + terminal_safeFall_fallTime;
             else
-                time = std::sqrt(2.0f * path_length / gravity);
+                time = sqrtf(2.0f * path_length / gravity);
         }
         else
         {
             if (path_length >= terminal_length)
                 time = (path_length - terminal_length) / terminalVelocity + terminal_fallTime;
             else
-                time = std::sqrt(2.0f * path_length / gravity);
+                time = sqrtf(2.0f * path_length / gravity);
         }
 
         return time;
@@ -76,7 +77,7 @@ namespace Movement
         {
             result = termVel * (t_passed - terminal_time) +
                 start_velocity * terminal_time +
-                gravity * terminal_time * terminal_time * 0.5f;
+                gravity * terminal_time * terminal_time*0.5f;
         }
         else
             result = t_passed * (start_velocity + t_passed * gravity * 0.5f);
